requirements.txt
python-telegram-bot>=21.0
emoji>=2.10.0
python-dotenv==1.0.0